Clean up create own files and shift things around
